IMAGE
Example screencapture of simulation at random timestep. Red x's represent robot position (with an additional history of 4 timesteps). Green markings are landmarks, the blue circle is the laser scanner range, and the green lines represent the robot 'sensing' the landmarks if in range of the sensor. The black dots are randomly distributed particles for particle filter localization